Split app components further
