+++ src/app/components/footer.tsx
@@ -1,0 +1,11 @@
+export function AppFooter() {
+    return (
+        <div className="centering-container app-footer-container">
+            <div className="app-footer">
+                <div>
+                    <i>Andydevs did this...</i>
+                </div>
+            </div>
+        </div>
+    )
+}
