+++ src/app/components/stats/index.tsx
@@ -1,0 +1,9 @@
+import { ElectronCount } from './count'
+
+export function ElectronStats() {
+    return (
+        <div className="stats-container centering-container">
+            <ElectronCount />
+        </div>
+    )
+}
